requirements.txt
streamlit==1.44.1
pandas==2.2.3
matplotlib==3.8.4
bcrypt==4.3.0
fpdf2==2.8.3
python-dotenv==1.0.1
pytz==2025.2
XlsxWriter==3.2.3
